+++ script/character/modifiers.ts
@@ -88,6 +88,84 @@ function getAllModifiers(char: Character | Player) {
   return modifiers;
 }
 
+function getModifierBreakdown(modif: string, char: Player | Enemy): any {
+  const breakdown = {
+    traitsP: 1,
+    perksP: 1,
+    statusesP: 1,
+    raceP: 1,
+    equipmentP: 1,
+    aspectP: 1,
+    classP: 1,
+  };
+  char.traits.forEach((trait: any) => {
+    if (trait.modifiers) {
+      Object.entries(trait.modifiers).forEach((modifier: any) => {
+        if (modifier[0] === modif && modifier[0].endsWith("P")) {
+          breakdown.traitsP *= 1 + modifier[1] / 100;
+        } else return;
+      });
+    }
+  });
+  char.perks?.forEach((perk: any) => {
+    perk.levels.forEach((level: any, lvl: number) => {
+      if (lvl >= perk.level) return;
+      if (level.modifiers) {
+        Object.entries(level.modifiers).forEach((modifier: any) => {
+          if (modifier[0] === modif && modifier[0].endsWith("P")) {
+            breakdown.perksP *= 1 + modifier[1] / 100;
+          }
+        });
+      }
+    });
+  });
+  char.statuses.forEach((status: any) => {
+    if (status.modifiers) {
+      Object.entries(status.modifiers).forEach((modifier: any) => {
+        if (modifier[0] === modif && modifier[0].endsWith("P")) {
+          breakdown.statusesP *= 1 + modifier[1] / 100;
+        }
+      });
+    }
+  });
+  if (char.race?.modifiers) {
+    Object.entries(char.race.modifiers).forEach((modifier: any) => {
+      if (modifier[0] === modif && modifier[0].endsWith("P")) {
+        breakdown.raceP *= 1 + modifier[1] / 100;
+      }
+    });
+  }
+  if (char.equipment) {
+    Object.values(char.equipment).forEach((item: any) => {
+      if (item?.modifiers) {
+        Object.entries(item.modifiers).forEach((modifier: any) => {
+          if (modifier[0] === modif && modifier[0].endsWith("P")) {
+            breakdown.equipmentP *= 1 + modifier[1] / 100;
+          }
+        });
+      }
+    });
+  }
+  if (char instanceof Player && typeof startingAspects !== "undefined") {
+    const aspectModifiers = startingAspects[char?.starting_aspect]?.modifiers;
+    if (aspectModifiers) {
+      Object.entries(aspectModifiers).forEach((modifier: any) => {
+        if (modifier[0] === modif && modifier[0].endsWith("P")) {
+          breakdown.aspectP *= 1 + modifier[1] / 100;
+        }
+      });
+    }
+  }
+  if (char instanceof Player && char.class?.modifiers) {
+    Object.entries(char.class.modifiers).forEach((modifier: any) => {
+      if (modifier[0] === modif && modifier[0].endsWith("P")) {
+        breakdown.classP *= 1 + modifier[1] / 100;
+      }
+    });
+  }
+  return breakdown;
+}
+
 function applyModifierToTotal(modifier: any, total: any) {
   const key = modifier[0];
   const value = modifier[1];